Fix test failures and 128-bit control word bugs
- Fix assembler cursor overflow for addresses at end of address space
  (use 32-bit cursor, only validate when emitting)
- Fix microcode decoder to use __uint128_t instead of uint64_t
  (controls with bit index >= 64 were silently dropped)
- Add decoder tests for high-bit controls to prevent regression
- Refactor LatchedWordRegister to capture PC value directly
  instead of going through bus (simpler and correct)
- Fix IPC latch test to match actual fetch preamble microcode
  (ipc.latch and pc.increment are never in the same step)
- Fix test helpers to use __uint128_t for control words
- Add timeout and max-cycles to asm integration tests

+++ microcode/include/irata2/microcode/debug/decoder.h
@@ -28,7 +28,7 @@ class MicrocodeDecoder {
    * @param control_word The encoded control word
    * @return Vector of control path strings for asserted bits
    */
-  std::vector<std::string> DecodeControlWord(uint64_t control_word) const;
+  std::vector<std::string> DecodeControlWord(__uint128_t control_word) const;
 
   /**
    * @brief Dump the entire program in readable format.
